Add files via upload
made it so to navigate calculator you type in text instead of numbers, added a bmi calculator, a gpa calculator, and fixed decision trees. just need to add code to do calculations. Note: if you use bmi calculator, just keep in mind it doesn't exactly mean whatever answer it gives you, it's a lil inaccurate anyways as a measure for describing 'healthiness'