Add files via upload
Change the directions so that you'd have to read the new readme attached to the calculator, as well as add loops to the random number generator and bmi calculators (for god knows what reason idk i was bored in class), if you're a programmer and want to see/edit the code feel free to; current to-do list is more quadratics and do an arraylist for statistics. other than those i have no other ideas for what I want to add, maybe something with calculus in the future.